refactor test tool
