+++ pyproject.toml
@@ -23,7 +23,13 @@ PyYAML = "^5.4.1"
 pylama = "^7.7.1"
 black = "^20.8b1"
 isort = "^5.7.0"
+pytest = "^6.2.2"
+pytest-isort = "^1.3.0"
+pytest-black = "^0.3.12"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+hcert = "hcert.cli:main"
